Moved WeightedPoints
